+++ openshift-client/src/main/java/io/fabric8/openshift/client/OpenShiftClient.java
@@ -52,6 +52,13 @@ public interface OpenShiftClient extends KubernetesClient {
   OpenShiftConsoleAPIGroupDSL console();
 
   /**
+   * API entrypoint for accessing OpenShift Cluster Operator resources (autoscaling.openshift.io)
+   *
+   * @return {@link OpenShiftClusterAutoscalingAPIGroupDSL} which contains respective resources in this API group
+   */
+  OpenShiftClusterAutoscalingAPIGroupDSL clusterAutoscaling();
+
+  /**
    * API entrypoint for accessing OpenShift operator APIGroup resources(operator.openshift.io/v1 and operator.openshift.io/v1alpha1)
    *
    * @return {@link OpenShiftOperatorAPIGroupDSL} which contains respective resources in this API group
@@ -91,6 +98,13 @@ public interface OpenShiftClient extends KubernetesClient {
    * @return {@link MachineConfigurationAPIGroupDSL} which contains respective resources in this ApiGroup
    */
   MachineConfigurationAPIGroupDSL machineConfigurations();
+
+  /*
+   * API entrypoint for accessing OpenShift Machine APIGroup resources(machine.openshift.io/v1beta1)
+   *
+   * @return {@link OpenShiftMachineAPIGroupDSL} which contains respective resources in this API group
+   */
+  OpenShiftMachineAPIGroupDSL machine();
 
   /**
    * API entrypoint for accessing OpenShift operator APIGroup resources(monitoring.coreos.com/v1)
@@ -237,6 +251,27 @@ public interface OpenShiftClient extends KubernetesClient {
   NonNamespaceOperation<OAuthClient, OAuthClientList, Resource<OAuthClient>> oAuthClients();
 
   /**
+   * API entrypoint for accessing PodSecurityPolicyReview (security.openshift.io/v1)
+   *
+   * @return InOutCreateable object for PodSecurityPolicyReview
+   */
+  OpenShiftCreateOnlyResourceOperationsImpl<PodSecurityPolicyReview, PodSecurityPolicyReview> podSecurityPolicyReviews();
+
+  /**
+   * API entrypoint for accessing PodSecurityPolicySelfSubjectReview (security.openshift.io/v1)
+   *
+   * @return InOutCreateable object for PodSecurityPolicySelfSubjectReview
+   */
+  OpenShiftCreateOnlyResourceOperationsImpl<PodSecurityPolicySelfSubjectReview, PodSecurityPolicySelfSubjectReview> podSecurityPolicySelfSubjectReviews();
+
+  /**
+   * API entrypoint for accessing PodSecurityPolicySubjectReview (security.openshift.io/v1)
+   *
+   * @return InOutCreateable object for PodSecurityPolicySubjectReview
+   */
+  OpenShiftCreateOnlyResourceOperationsImpl<PodSecurityPolicySubjectReview, PodSecurityPolicySubjectReview> podSecurityPolicySubjectReviews();
+
+  /**
    * API entrypoint for accessing Project operations(project.openshift.io/v1)
    *
    * @return {@link ProjectOperation} for Project specific operations
@@ -294,6 +329,13 @@ public interface OpenShiftClient extends KubernetesClient {
   MixedOperation<TemplateInstance, TemplateInstanceList, Resource<TemplateInstance>> templateInstances();
 
   /**
+   * API entrypoint for OpenShift Node Tuning operator model(tuned.openshift.io/v1)
+   *
+   * @return {@link OpenShiftTunedAPIGroupDSL} for access to various available resources
+   */
+  OpenShiftTunedAPIGroupDSL tuned();
+
+  /**
    * API entrypoint for BrokerTemplateInstance(template.openshift.io/v1)
    *
    * @return {@link NonNamespaceOperation} for BrokerTemplateInstance
@@ -341,33 +383,33 @@ public interface OpenShiftClient extends KubernetesClient {
    * API entrypoint for LocalSubjectAccessReview (authorization.openshift.io/v1)
    * This only supports create operation. SubjectAccessReviewResponse from server is returned as output
    *
-   * @return {@link OpenShiftLocalSubjectAccessReviewOperationsImpl} for LocalSubjectAccessReview
+   * @return {@link OpenShiftCreateOnlyResourceOperationsImpl} for LocalSubjectAccessReview
    */
-  OpenShiftLocalSubjectAccessReviewOperationsImpl<LocalSubjectAccessReview, SubjectAccessReviewResponse> localSubjectAccessReviews();
+  OpenShiftCreateOnlyResourceOperationsImpl<LocalSubjectAccessReview, SubjectAccessReviewResponse> localSubjectAccessReviews();
 
   /**
    * API entrypoint for LocalResourceAccessReview (authorization.openshift.io/v1)
    * This only supports create operation. ResourceAccessReviewResponse from server is returned as output
    *
-   * @return {@link OpenShiftLocalSubjectAccessReviewOperationsImpl} for LocalResourceAccessReview
+   * @return {@link OpenShiftCreateOnlyResourceOperationsImpl} for LocalResourceAccessReview
    */
-  OpenShiftLocalSubjectAccessReviewOperationsImpl<LocalResourceAccessReview, ResourceAccessReviewResponse> localResourceAccessReviews();
+  OpenShiftCreateOnlyResourceOperationsImpl<LocalResourceAccessReview, ResourceAccessReviewResponse> localResourceAccessReviews();
 
   /**
    * API entrypoint for SelfSubjectRulesReview (authorization.openshift.io/v1)
    * This only supports create operation. SelfSubjectRulesReview from server is returned as output
    *
-   * @return {@link OpenShiftLocalSubjectAccessReviewOperationsImpl} for SelfSubjectRulesReview
+   * @return {@link OpenShiftCreateOnlyResourceOperationsImpl} for SelfSubjectRulesReview
    */
-  OpenShiftLocalSubjectAccessReviewOperationsImpl<SelfSubjectRulesReview, SelfSubjectRulesReview> selfSubjectRulesReviews();
+  OpenShiftCreateOnlyResourceOperationsImpl<SelfSubjectRulesReview, SelfSubjectRulesReview> selfSubjectRulesReviews();
 
   /**
    * API entrypoint for SubjectRulesReview (authorization.openshift.io/v1)
    * This only supports create operation. SubjectRulesReview from server is returned as output
    *
-   * @return {@link OpenShiftLocalSubjectAccessReviewOperationsImpl} for SubjectRulesReview
+   * @return {@link OpenShiftCreateOnlyResourceOperationsImpl} for SubjectRulesReview
    */
-  OpenShiftLocalSubjectAccessReviewOperationsImpl<SubjectRulesReview, SubjectRulesReview> subjectRulesReviews();
+  OpenShiftCreateOnlyResourceOperationsImpl<SubjectRulesReview, SubjectRulesReview> subjectRulesReviews();
 
   /**
    * API entrypoint for ClusterRole (authorization.openshift.io/v1)