+++ openshift-client/src/main/java/io/fabric8/openshift/client/internal/OpenShiftClusterOperationsImpl.java
@@ -53,7 +53,7 @@ public class OpenShiftClusterOperationsImpl extends ClusterOperationsImpl {
   }
 
   private VersionInfo fetchOpenshift4Version() throws IOException, ParseException {
-    Response response = handleVersionGet(OPENSHIFT4_VERSION_ENDPOINT);
+    Response response = handleVersionGet(getRequest(OPENSHIFT4_VERSION_ENDPOINT));
     if (response.isSuccessful() && response.body() != null) {
       ClusterVersionList clusterVersionList = Serialization.jsonMapper().readValue(response.body().string(), ClusterVersionList.class);
       if (!clusterVersionList.getItems().isEmpty()) {
@@ -66,7 +66,7 @@ public class OpenShiftClusterOperationsImpl extends ClusterOperationsImpl {
   private static VersionInfo parseVersionInfoFromClusterVersion(ClusterVersion clusterVersion) throws ParseException {
     String[] versionParts = clusterVersion.getStatus().getDesired().getVersion().split("\\.");
     VersionInfo.Builder versionInfoBuilder = new VersionInfo.Builder();
-    if (versionParts.length == 3) {
+    if (versionParts.length > 2) {
       versionInfoBuilder.withMajor(versionParts[0]);
       versionInfoBuilder.withMinor(versionParts[1] + "." + versionParts[2]);
     }